+++ pyvergeos/resources/vm_recipes.py
@@ -156,7 +156,7 @@ class VmRecipe(ResourceObject):
     @property
     def instance_count(self) -> int:
         """Get the number of deployed instances."""
-        return int(self.get("instances", 0))
+        return int(self.require_projected("instances"))
 
     @property
     def instances(self) -> VmRecipeInstanceManager: